Added async configuration and extracted room operations into their own controller
