+++ src/main/kotlin/com/benbasten/groupmusicserver/controller/MusicController.kt
@@ -30,32 +30,6 @@ class MusicController(private val musicService: MusicService, private val roomSe
         return musicService.getTrackListings()
     }
 
-    @GetMapping("/room/now-playing")
-    fun getRoomNowPlaying(@RequestParam("roomId") roomId: Int): ResponseEntity<StreamingResponseBody> {
-        val file: File = roomService.getCurrentTrack(roomId) ?: return ResponseEntity.noContent().build()
-        val responseBody = StreamingResponseBody { outputStream: OutputStream -> Files.copy(file.toPath(), outputStream) }
-        return ResponseEntity.ok()
-            .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=${file.name}")
-            .header("Cache-Control", "no-store")
-            .contentType(MediaType.asMediaType(MimeType("audio", "mpeg")))
-            .body(responseBody)
-    }
-
-    @GetMapping("/room/queue")
-    fun getRoomQueue(@RequestHeader("roomId") roomId: Int): List<Track> {
-        return roomService.getQueue(roomId)
-    }
-
-    @PostMapping("/room/queue/add")
-    fun addToQueue(@RequestBody trackId: Int, @RequestHeader("roomId") roomId: Int): List<Track> {
-        return roomService.addToQueue(roomId, trackId)
-    }
-
-    @PostMapping("/room/next-track")
-    fun nextTrack(@RequestHeader("roomId") roomId: Int): List<Track> {
-        return roomService.nextTrack(roomId)
-    }
-
     @PostMapping("/create")
     fun createRoom(): Room {
         return roomService.createRoom()
